+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupListTable.cs
@@ -129,7 +129,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GSub
             GSubTable table,
             GlyphSubstitutionCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count)
         {
             foreach (LookupSubTable subTable in this.LookupSubTables)
@@ -171,7 +171,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GSub
             GSubTable table,
             GlyphSubstitutionCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count);
     }
 }
